Split combobox and select styles to separate files
