+++ src/common/components/combobox/combobox.module.scss
@@ -1,0 +1,10 @@
+.combobox {
+  > div:focus-within {
+    box-shadow: none;
+
+    input:global(.focus-visible) {
+      outline: var(--focus-outline-width) solid var(--focus-outline-color);
+      outline-offset: 1px;
+    }
+  }
+}
