+++ src/common/components/select/select.module.scss
@@ -1,30 +1,8 @@
-.combobox {
-  > div:focus-within {
-    box-shadow: none;
-
-    input:global(.focus-visible) {
-      outline: var(--focus-outline-width) solid var(--focus-outline-color);
-      outline-offset: 1px;
-    }
-  }
-}
-
 .label {
   .text {
     color: var(--color-black-90);
   }
   .amount {
     color: var(--color-black-60);
-  }
-}
-
-.select {
-  > div:focus-within {
-    box-shadow: none;
-
-    button:global(.focus-visible) {
-      outline: var(--focus-outline-width) solid var(--focus-outline-color);
-      outline-offset: 1px;
-    }
   }
 }
