+++ src/common/components/singleSelect/singleSelect.module.scss
@@ -1,0 +1,10 @@
+.select {
+  > div:focus-within {
+    box-shadow: none;
+
+    button:global(.focus-visible) {
+      outline: var(--focus-outline-width) solid var(--focus-outline-color);
+      outline-offset: 1px;
+    }
+  }
+}
